+++ include/util/SingleBuffer.inl
