+++ src/util/TimeSeries.cpp
@@ -8,7 +8,7 @@
 // with this source code package.
 //
 
-#include "TimeSeries.hpp"
+#include "util/TimeSeries.hpp"
 
 
 using Json = nlohmann::json;
